+++ yarn-project/circuit-types/src/tx/processed_tx.ts
@@ -27,6 +27,7 @@ import {
   type TUBE_PROOF_LENGTH,
   type VerificationKeyData,
 } from '@aztec/circuits.js';
+import { siloL2ToL1Message } from '@aztec/circuits.js/hash';
 
 import { type CircuitName } from '../stats/stats.js';
 
@@ -267,12 +268,10 @@ export function toTxEffect(tx: ProcessedTx, gasFees: GasFees): TxEffect {
     tx.data.end.noteHashes.filter(h => !h.isZero()),
     tx.data.end.nullifiers.filter(h => !h.isZero()),
     tx.data.end.l2ToL1Msgs
-      .filter(h => !h.isEmpty())
       .map(message =>
-        Fr.fromBuffer(
-          message.silo(tx.data.constants.globalVariables.version, tx.data.constants.globalVariables.chainId),
-        ),
-      ),
+        siloL2ToL1Message(message, tx.data.constants.txContext.version, tx.data.constants.txContext.chainId),
+      )
+      .filter(h => !h.isZero()),
     tx.finalPublicDataUpdateRequests.map(t => new PublicDataWrite(t.leafSlot, t.newValue)).filter(h => !h.isEmpty()),
     tx.data.end.noteEncryptedLogPreimagesLength,
     tx.data.end.encryptedLogPreimagesLength,
