+++ yarn-project/circuits.js/src/hash/hash.test.ts
@@ -1,7 +1,7 @@
 import { times } from '@aztec/foundation/collection';
-import { setupCustomSnapshotSerializers } from '@aztec/foundation/testing';
+import { setupCustomSnapshotSerializers, updateInlineTestData } from '@aztec/foundation/testing';
 
-import { AztecAddress, Fr, MAX_ARGS_LENGTH } from '../index.js';
+import { AztecAddress, EthAddress, Fr, L2ToL1Message, MAX_ARGS_LENGTH, ScopedL2ToL1Message } from '../index.js';
 import { makeAztecAddress } from '../tests/factories.js';
 import {
   computeNoteHashNonce,
@@ -10,6 +10,7 @@ import {
   computeSecretHash,
   computeUniqueNoteHash,
   computeVarArgsHash,
+  siloL2ToL1Message,
   siloNoteHash,
   siloNullifier,
 } from './hash.js';
@@ -88,5 +89,28 @@ describe('hash', () => {
 
     // Value used in "compute_var_args_hash" test in hash.nr
     // console.log("hash", hash);
+  });
+
+  it('L2ToL1Message siloing matches Noir', () => {
+    const version = new Fr(4);
+    const chainId = new Fr(5);
+
+    const nonEmpty = new ScopedL2ToL1Message(
+      new L2ToL1Message(EthAddress.fromField(new Fr(1)), new Fr(2), 0),
+      AztecAddress.fromField(new Fr(3)),
+    );
+
+    const nonEmptyHash = new Fr(siloL2ToL1Message(nonEmpty, version, chainId));
+
+    expect(nonEmptyHash.toString()).toMatchInlineSnapshot(
+      `"0x00c6155d69febb9d5039b374dd4f77bf57b7c881709aa524a18acaa0bd57476a"`,
+    );
+
+    // Run with AZTEC_GENERATE_TEST_DATA=1 to update noir test data
+    updateInlineTestData(
+      'noir-projects/noir-protocol-circuits/crates/types/src/hash.nr',
+      'hash_from_typescript',
+      nonEmptyHash.toString(),
+    );
   });
 });
